Renamed datasetCreator to datasetPublisher
