+++ src/main/java/nl/knaw/dans/ingest/config/DataverseExtra.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.ingest.core.config;
+package nl.knaw.dans.ingest.config;
 
 public class DataverseExtra {
     private int publishAwaitUnlockMaxRetries;
