+++ src/test/java/nl/knaw/dans/ingest/core/service/mapper/MappingTestHelper.java
@@ -21,9 +21,9 @@ import io.dropwizard.configuration.ConfigurationException;
 import io.dropwizard.configuration.YamlConfigurationFactory;
 import io.dropwizard.jackson.Jackson;
 import io.dropwizard.jersey.validation.Validators;
-import nl.knaw.dans.ingest.DdIngestFlowConfiguration;
-import nl.knaw.dans.ingest.IngestFlowConfigReader;
-import nl.knaw.dans.ingest.core.config.IngestFlowConfig;
+import nl.knaw.dans.ingest.config.DdIngestFlowConfiguration;
+import nl.knaw.dans.ingest.config.IngestFlowConfigReader;
+import nl.knaw.dans.ingest.config.IngestFlowConfig;
 import nl.knaw.dans.ingest.core.domain.VaultMetadata;
 import nl.knaw.dans.ingest.core.service.XmlReaderImpl;
 import nl.knaw.dans.lib.dataverse.model.dataset.CompoundMultiValueField;
